refactor: remove and restructure scripts and docs in vscode extension
Co-developed-by: Aone Copilot <noreply@alibaba-inc.com>
