+++ extensions/vscode/README.md
@@ -2,8 +2,6 @@
 
 基于 [`open-code-review`](https://www.npmjs.com/package/@alibaba-group/open-code-review) (`ocr`) CLI 的 VSCode 代码审查插件。以 Preact WebView 还原原型交互体验，把 AI 代码审查能力集成进编辑器：在侧边栏发起审查、流式查看日志、在编辑器内逐条应用/忽略/标记误报评论，并与侧边栏双向同步。
 
-本插件可作为**独立扩展**运行，也能通过一键脚本同步复用到 `aone-copilot-vscode`。
-
 ---
 
 ## 功能
@@ -66,28 +64,6 @@ yarn build        # 生产构建（webpack production）
 ```
 
 产物：`out/extension.js`（Extension Host）+ `out/webview.js`（WebView SPA）。
-
----
-
-## 复用到 aone-copilot-vscode
-
-把核心源码同步进 aone 的 `src/ocr/`，并幂等 merge `contributes` 到 aone 的 `package.json`：
-
-```bash
-yarn sync:aone --target ../aone-copilot-vscode/src/ocr
-```
-
-脚本会：
-
-1. 复制 `src/`（排除独立入口 `extension.ts` 与测试文件）到目标目录。
-2. 幂等 merge 本插件的 `contributes`（命令 / 视图 / 视图容器 / 评论菜单）进 aone 的 `package.json`，重复执行不会产生重复条目，并会清理旧的 `ocr.*` 条目。
-3. 写入 `.synced-version` 记录来源 commit 与时间。
-
-aone 侧需做的适配（量很小）：
-
-1. **调用入口**：在 aone 的 `src/extension.ts` 中调用复制进去的 `activateOcr(context, { ... })`（来自 `src/ocr/extension/index.ts`），并在停用时调用 `deactivateOcr()`。
-2. **WebView 构建**：在 aone 的 webpack 增加一个入口指向 `src/ocr/webview/index.tsx`，产物命名为 `webview.js` 放到 aone 的 `out/`，使 `SidebarProvider` 的 `asWebviewUri('out/webview.js')` 能命中。
-3. **contributes merge**：已由同步脚本自动完成，无需手工编辑。
 
 ---
